Fix an issue with generating permalinks for some sections
Permalink’s are created for every section header. These permalinks include the ID of their parent header, but if there is no parent header (e.g. no h2 for an h3), the code would throw an error. This caused issues like the legend and tooltips not working for the framework comparison matrix.