Launch donejs-cli from Node
Instead of calling out to node_modules/.bin/donejs we are going to call out to "node node_modules/donejs-cli/bin/donejs". This is only for donejs init.

The reason is that on windows by calling out to donejs.cmd npm is unable to remove this file when it reinstalls donejs-cli.