Replace the Upcoming Events BitHub embed with Google Calendar integration
The site now uses Google Calendar for the Upcoming Events section of the Community page. Any non-recurring events in the DoneJS calendar will be displayed.

I renamed the components/controls and changed the functionality, but I didn’t spend much time refactoring the code, adding tests, etc. If that’s required before merging this in, I can spend more time on it.

`moment-timezone` is required as a new dependency because the Google Calendar API provides the `dateTime` in the calendar’s default timezone (Chicago) and also includes the event timezone. Because there will be mismatches between the calendar’s default timezone and the events’ timezones, we need to calculate the event’s real time based on the difference between Chicago and the event’s timezone.

Fixes #681