Make donejs upgrade upgrade all can deps
This makes it so that running `donejs upgrade` it will check the
compatible version of `can`'s package.json and update any depenencies
that you have therein.

This means if you depend on `can-stache` you'll get the correctly
upgraded version of can-stache. This is true for ecosystem packages as
well.

Closes #1132 and #1092